maintain: remove apimigrator
