+++ internal/server/handlers.go
@@ -23,7 +23,6 @@ import (
 type API struct {
 	t          *Telemetry
 	server     *Server
-	migrations []apiMigration
 	openAPIDoc openapi3.T
 }
